% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pjaehee99-package.R
\docType{package}
\name{hw04pjaehee99-package}
\alias{hw04pjaehee99}
\alias{hw04pjaehee99-package}
\title{hw04pjaehee99: STAT-413_Homework Creating a Package}
\description{
What the package does (Two functions inside this package).
}
\author{
\strong{Maintainer}: Jaehee Lee \email{jl8883a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
