% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n.R
\name{myseq_n}
\alias{myseq_n}
\title{Function 1 Recursive Sequence}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{x vector}

\item{n}{n num}
}
\value{
This function returns element n
}
\description{
Function 1 Recursive Sequence
}
\examples{
myseq_n(x=c(2,4,3), n= 4)
}
